Solved merge conflicts an applied minor fixes
