+++ app/src/main/AndroidManifest.xml
@@ -11,15 +11,18 @@
         android:theme="@style/Theme.Autoshare">
         <activity
             android:name="com.riteh.autoshare.ui.auth.AuthActivity"
-            android:exported="false" />
-        <activity
-            android:name=".MainActivity"
-            android:exported="true" >
+            android:exported="true">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".ui.home.MainActivity"
+            android:exported="false">
+
         </activity>
     </application>
