--- app/src/main/java/com/riteh/autoshare/model/SearchDataModel.kt
@@ -1,3 +1,0 @@
-package com.riteh.autoshare.model
-
-data class SearchDataModel(var textForUI: String)
