+++ app/src/main/java/com/riteh/autoshare/ui/base/ViewModelFactory.kt
@@ -11,7 +11,7 @@ class ViewModelFactory(
     private val repository: BaseRepository
 ) : ViewModelProvider.NewInstanceFactory() {
 
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return when{
             modelClass.isAssignableFrom(AuthViewModel::class.java) -> AuthViewModel(repository as AuthRepository) as T
