+++ app/src/main/java/com/riteh/autoshare/ui/home/InfoFragment.kt
@@ -1,4 +1,4 @@
-package com.riteh.autoshare.view
+package com.riteh.autoshare.ui.home
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -7,26 +7,25 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.riteh.autoshare.R
-import com.riteh.autoshare.viewmodel.SearchViewModel
 
-class SearchFragment : Fragment() {
+class InfoFragment : Fragment() {
 
     companion object {
-        fun newInstance() = SearchFragment()
+        fun newInstance() = InfoFragment()
     }
 
-    private lateinit var viewModel: SearchViewModel
+    private lateinit var viewModel: InfoViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.search_fragment, container, false)
+        return inflater.inflate(R.layout.info_fragment, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        viewModel = ViewModelProvider(this)[SearchViewModel::class.java]
+        viewModel = ViewModelProvider(this)[InfoViewModel::class.java]
     }
 
 }
