+++ app/src/main/java/com/riteh/autoshare/ui/home/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.riteh.autoshare.view
+package com.riteh.autoshare.ui.home
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
