+++ app/src/main/java/com/riteh/autoshare/ui/home/SearchViewModel.kt
@@ -1,0 +1,7 @@
+package com.riteh.autoshare.ui.home
+
+import androidx.lifecycle.ViewModel
+
+class SearchViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
