+++ app/src/main/java/com/riteh/autoshare/ui/home/UserFragment.kt
@@ -1,4 +1,4 @@
-package com.riteh.autoshare.view
+package com.riteh.autoshare.ui.home
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -7,27 +7,25 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.riteh.autoshare.R
-import com.riteh.autoshare.viewmodel.AddViewModel
-import com.riteh.autoshare.viewmodel.SearchViewModel
 
-class AddFragment : Fragment() {
+class UserFragment : Fragment() {
 
     companion object {
-        fun newInstance() = AddFragment()
+        fun newInstance() = UserFragment()
     }
 
-    private lateinit var viewModel: AddViewModel
+    private lateinit var viewModel: UserViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.add_fragment, container, false)
+        return inflater.inflate(R.layout.user_fragment, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        viewModel = ViewModelProvider(this)[AddViewModel::class.java]
+        viewModel = ViewModelProvider(this)[UserViewModel::class.java]
     }
 
 }
