--- app/src/main/java/com/riteh/autoshare/viewmodel/MessagesViewModel.kt
@@ -1,7 +1,0 @@
-package com.riteh.autoshare.viewmodel
-
-import androidx.lifecycle.ViewModel
-
-class MessagesViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
