+++ app/src/main/res/layout/add_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.AddFragment">
+    tools:context=".ui.home.AddFragment">
 
     <TextView
         android:layout_width="match_parent"
