+++ app/src/main/res/layout/info_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.InfoFragment">
+    tools:context=".ui.home.InfoFragment">
 
     <TextView
         android:layout_width="match_parent"
