+++ app/src/main/res/layout/search_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SearchFragment">
+    tools:context=".ui.home.SearchFragment">
 
     <TextView
         android:layout_width="match_parent"
