+++ app/src/main/res/layout/user_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.UserFragment">
+    tools:context=".ui.home.UserFragment">
 
     <TextView
         android:layout_width="match_parent"
